# openapi/admin-bff-openapi-v1.yaml
openapi: 3.0.3
info:
  title: Admin BFF API
  version: 1.0.0
  description: |
    Admin BFF (screen-first) API for Admin Platform.
    - Aggregates Core Domain Services for UI use-cases
    - Applies RBAC/ABAC and context enforcement (tenant/org/cycle)
servers:
  - url: https://api.example.com/admin-bff/v1
    description: Production
  - url: https://staging-api.example.com/admin-bff/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /approvals:
    get:
      tags: [Approvals]
      summary: List approval requests (queue)
      operationId: listApprovals
      parameters:
        - $ref: "#/components/parameters/HeaderTenantId"
        - $ref: "#/components/parameters/HeaderOrgId"
        - $ref: "#/components/parameters/HeaderCycleId"
        - $ref: "#/components/parameters/HeaderRequestId"
        - $ref: "#/components/parameters/HeaderAcceptLanguage"
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApprovalState"
        - name: type
          in: query
          required: false
          schema:
            type: string
          example: MEMBER_ONBOARDING
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, URGENT]
        - $ref: "#/components/parameters/QueryQ"
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QuerySize"
        - $ref: "#/components/parameters/QuerySort"
      responses:
        "200":
          description: Approval queue
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedApprovalsResponse"
        "304":
          description: Not modified
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /approvals/{approvalId}:
    get:
      tags: [Approvals]
      summary: Get approval request detail
      operationId: getApproval
      parameters:
        - $ref: "#/components/parameters/HeaderTenantId"
        - $ref: "#/components/parameters/HeaderOrgId"
        - $ref: "#/components/parameters/HeaderCycleId"
        - $ref: "#/components/parameters/HeaderRequestId"
        - $ref: "#/components/parameters/HeaderAcceptLanguage"
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
          example: apr_01J9Z0JZ2K1P0
      responses:
        "200":
          description: Approval detail
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /approvals/{approvalId}/decide:
    post:
      tags: [Approvals]
      summary: Decide an approval request (approve/reject/request changes)
      operationId: decideApproval
      parameters:
        - $ref: "#/components/parameters/HeaderTenantId"
        - $ref: "#/components/parameters/HeaderOrgId"
        - $ref: "#/components/parameters/HeaderCycleId"
        - $ref: "#/components/parameters/HeaderRequestId"
        - $ref: "#/components/parameters/HeaderAcceptLanguage"
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
          example: apr_01J9Z0JZ2K1P0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Decision applied
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /approvals/bulk-decide:
    post:
      tags: [Approvals]
      summary: Bulk decide approvals (with per-item results)
      operationId: bulkDecideApprovals
      parameters:
        - $ref: "#/components/parameters/HeaderTenantId"
        - $ref: "#/components/parameters/HeaderOrgId"
        - $ref: "#/components/parameters/HeaderCycleId"
        - $ref: "#/components/parameters/HeaderRequestId"
        - $ref: "#/components/parameters/HeaderAcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkApprovalDecisionRequest"
      responses:
        "200":
          description: Bulk decision results
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkApprovalDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    XRequestId:
      schema:
        type: string
      description: Correlation id for the request
    ETag:
      schema:
        type: string
      description: Entity tag for cache validation (If-None-Match)
  parameters:
    HeaderTenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
      description: Tenant identifier
    HeaderOrgId:
      name: X-Org-Id
      in: header
      required: true
      schema:
        type: string
      description: Organization identifier
    HeaderCycleId:
      name: X-Cycle-Id
      in: header
      required: false
      schema:
        type: string
      description: Cycle identifier (required for cycle-scoped operations)
    HeaderRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Client correlation id
    HeaderAcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        default: en
        enum: [pt-BR, en, es]
      description: Response language
...
    QueryQ:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Free text search
...
    QueryPage:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    QuerySize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    QuerySort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: Sort expression (e.g. createdAt:desc,name:asc)
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized (missing/invalid token)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict (state/version/business constraints)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - requestId
        -  timestamp
      properties:
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: "Invalid filter: status"
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        requestId:
          type: string
          example: 8ad1f3b1-0f0f-4c3a-a0e2-5d1d9a0f8a1c
        timestamp:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: status
        code:
          type: string
          example: INVALID_ENUM
        message:
          type: string
          example: Must be one of ACTIVE, SUSPENDED, PENDING
    PagedMeta:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 25
        totalItems:
          type: integer
          example: 257
        totalPages:
          type: integer
          example: 11
...
    ApprovalState:
      type: string
      enum: [PENDING, IN_REVIEW, APPROVED, REJECTED, CHANGES_REQUESTED, CANCELED]
    ApprovalSummary:
      type: object
      required: [approvalId, type, state, entityType, entityId, createdAt]
      properties:
        approvalId:
          type: string
        type:
          type: string
        state:
          $ref: "#/components/schemas/ApprovalState"
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
          nullable: true
        entityType:
          type: string
          example: Member
        entityId:
          type: string
        requester:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PagedApprovalsResponse:
      type: object
      required: [meta, items]
      properties:
        meta:
          $ref: "#/components/schemas/PagedMeta"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalSummary"
    ApprovalDetailResponse:
      type: object
      required: [approval]
      properties:
        approval:
          $ref: "#/components/schemas/ApprovalDetail"
    ApprovalDetail:
      type: object
      required: [approvalId, type, state, entityType, entityId, createdAt]
      properties:
        approvalId:
          type: string
        type:
          type: string
        state:
          $ref: "#/components/schemas/ApprovalState"
        entityType:
          type: string
        entityId:
          type: string
        payload:
          type: object
          additionalProperties: true
          description: Domain payload snapshot for UI rendering
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
        createdAt:
          type: string
          format: date-time
    Decision:
      type: object
      required: [decision, decidedBy, decidedAt]
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT, REQUEST_CHANGES]
        reasonCode:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        decidedBy:
          type: string
        decidedAt:
          type: string
          format: date-time
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT, REQUEST_CHANGES]
        reasonCode:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
          maxLength: 1500
        expectedState:
          $ref: "#/components/schemas/ApprovalState"
    ApprovalDecisionResponse:
      type: object
      required: [approvalId, previousState, newState, decidedAt]
      properties:
        approvalId:
          type: string
        previousState:
          $ref: "#/components/schemas/ApprovalState"
        newState:
          $ref: "#/components/schemas/ApprovalState"
        decidedAt:
          type: string
          format: date-time
    BulkApprovalDecisionRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BulkApprovalDecisionItem"
    BulkApprovalDecisionItem:
      type: object
      required: [approvalId, decision]
      properties:
        approvalId:
          type: string
        decision:
          type: string
          enum: [APPROVE, REJECT, REQUEST_CHANGES]
        reasonCode:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        expectedState:
          $ref: "#/components/schemas/ApprovalState"
    BulkApprovalDecisionResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BulkDecisionResult"
    BulkDecisionResult:
      type: object
      required: [approvalId, success]
      properties:
        approvalId:
          type: string
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ErrorResponse"
...